Add test coverage for the TraceStore and fix bug uncovered in enqueueTrace function (#7)
Here, we add tests that cover creating a new TraceStore, adding spans to
the store, and handling the store's map and queue when the number of
traceIDs exceeds the preset maximum.

During this process, a bug was found in the enqueueTrace function that
caused traces to be deleted prematurely. This was resolved by moving the
queue length check to the else clause that handles adding **new**
traceIDs to the map/queue.