First Pass at Ordering Spans for Display in Waterfall View (#50)
Spans are ordered according to their parent/child relationships.
Additionally, spans that share a `parentSpanID` are sorted by earliest
start time.

This is being done in `TraceView` rather than `WaterfallView` because
that is where we set the `selectedSpanID` when changing routes. It was
simpler to pull the whole thing up and pass the sorted array to
`WaterfallView`.

Currently, orphan spans and incomplete trees aren't handled. I will do
this thing in my next PR.

Also, the depth is visually represented with padding, which is not very
clear. I will find a better solution once I have taken care of the
orphans. I'm aware that I sound like a supervillain. This is fine.


![OrderedSpansPass1](https://user-images.githubusercontent.com/56372758/210431385-29b4cc73-9822-4383-ab4c-25412729910e.jpg)