update sidebar view to reflect new summary data (#57)
- Made a new UI type for trace summary in order to discriminate between
summaries where the root span is present, and incomplete traces where it
is missing.
- Switched from fixed to variable sized list in `react-window` to
account for the difference in height between a complete and partial
trace summary.
- Reimplemented basically the whole `trace-list`, or at least enough
that git decided not to track individual changes. It looks better now.
- Made a sidebar-view folder so sidebar components are no longer
rattling around loose. That's what the import updates are about. Also, I
renamed stuff. It makes more sense to me now.
- Fixed the bug where the sidebar would squish with `flexShrink="0"`,
which is minor but important.
- LOOK AT IT:

![sidebarDay](https://user-images.githubusercontent.com/56372758/212794983-6e8fe954-f009-482a-9556-a97d8e994de2.jpg)

![SidebarNight](https://user-images.githubusercontent.com/56372758/212795005-992eb0e9-c9f9-44ee-a2b1-03fee870ca4f.jpg)