build collector docker image
This change adds a distribution directory which contains the Dockerfile,
manifest, and default configuration. It also adds a make target to build
a custom collector, which includes the desktop exporter. This custom
collector is then used to build and publish a docker image.

The internal/tools directory will be used for tools like the collector builder and any tools used to generate telemetry.

Signed-off-by: Alex Boten <alex@boten.ca>