add support for metrics and logs
This change adds support for metrics & logs to the exporter code. The
commits are as follows:

1. add methods to the factory to make the component available for
   metrics/logs pipelines
2. add extract methods for logs/metrics

This is also adding a telemetry store alongside the existing trace store. The
idea being the current functionality will continue to work while the
telemetry store is being filled in. In this PR:

- created the `internal/telemetry` package and ended up moving a big
  chunk of the existing models in there
- added a telemetry store to the server & exporter
- adding code to generate samples for logs/metrics
- adding endpoints /telemetry & /telemetry/{id}

There's still many details to fill in, but I wanted to get the pieces
in place to fill in the details later, without too big of a PR.

Signed-off-by: Alex Boten <alex@boten.ca>