Add keyboard navigation to spans in waterfall view (#115)
The spans in the waterfall view can now be traversed using the up and
down arrows as well as the j and k vim equivalents.
Placeholders for missing spans are automatically skipped. The span's
data is automatically displayed in the detail view.

Note: I disabled the default behaviour of the keys because arrow presses
scrolled through whatever list you happen to have left your cursor in.
Right now, it won't scroll as you move up and down the list. I will sync
it up with `react-window`'s scroll to item in the next PR, but wanted to
get this logic out of the way first.

Look at it go!

![KeyboardNavSpans](https://user-images.githubusercontent.com/56372758/236705277-82b757fc-0ff1-4215-8157-9c6c3bbad19d.gif)