Add host flag (#129)
Not a go expert, so maybe the implementation needs to look differently.

I wanted to run otel-desktop-viewer in a docker container but didn't get
it working because it only listened on localhost. Of course I could
build a custom collector and set the config, but I thought this way it
is more convenient.