WIP: Desktop viewer 2.0 - Clean Slate (#147)
- [x] Delete everything
- [x] Spin up an updated Collector
- [x] Fix main to create the config we need from command line
- [x] Put the Makefile back in
- [x] Refactor the code to separate the collector and the exporter
- [x] Refactor the contents of the `telemetry` folder, getting rid of
`payload_processor.go`
- [x] Move `server.go to its own folder, along with the `static`
directory, and fix the server to work with the new telemetry methods
- [x] Update the sample data generator and move it into telemetry 
- [x] Fix `factory.go` to work with the new version of OTel
- [x] Fix `exporter.go` to work with the new version of OTel
- [x] Put the front end back in  

![A plump, fluffy capybara holding a mop in its tiny paws, busily
engaged in spring cleaning. Around it, you can see a bucket filled with
soapy water and some cleaning supplies. The capybara is working
diligently, creating a picture of cleanliness and order as it moves
around. The mop is swishing back and forth, displacing a trail of
bubbles that capture the light, creating rainbows. The capybara itself
is a brown, fluffy creature, large for a rodent but gentle and earnest
in its task of tidying
up.](https://github.com/user-attachments/assets/d14132f5-8a58-4d3c-a663-8b97b5bd6c74)