Allow running in Docker (#168)
I run my dev environment using Docker, and don't want to/can't easily
install stuff on my computer with Homebrew.

The `Dockerfile` in this PR worked for me, but I want to contribute
whatever makes sense to this project if others want to do the same.

I'm not sure the best way, but I can think of a few scenarios:

1. **Publish a Docker image others can use.** This would require a
`Dockerfile` similar to what is in this PR as well as necessary
scripts and credentials to build and push the image. It would also
require this repo's owners to maintain that image.
2. **Just provide instructions.** This would involve adding something to
the `README` I guess?

I'm happy to do more work along either option, so if one is prefered
over the other, I will have more questions on how to properly
proceed.