Duckdb 1.2 update (#172)
- Not gonna lie, I updated everything under the sun trying to tack down
the Weird Mystery Error, including go and the otel collector. Changes to
`desktopcollector/components.go` and `desktopexporter/factory.go`
reflect the new version.
- I took `go.work` and `go.work.sum` out of `.gitignore` because
workspaces are very handy and forgetting they are a thing can cause no
end of grief.
- The actual work was updating to the next version of DuckDb. I
leveraged DuckDb's ability to read/write json to deal with data types
that we don't quite have support for in go-duckdb. I got to delete a
whole bunch of parsing logic, and also got parametrized inserts back!
It's an improvement!


![Gemini_Generated_Image_cp1ayucp1ayucp1a](https://github.com/user-attachments/assets/d08831aa-f031-4b7f-ac0c-823b0b21043b)