Feat/logs back end (#181)
Hi! Hello! This PR enables us to receive logs, store them in duckDB, and
get them to our front-end.

I also changed how we handle timestamps because the way I was doing it
before was not well thought out and became a problem for falling back to
observed timestamp for sorting logs when a record's timestamp was
unavailable. Anyway, timestamps are now being stored as
[UnixNano](https://pkg.go.dev/time#Time.UnixNano), and is marshalled to
JSON like this to avoid losing nanosecond precision:
`
// PreciseTimestamp represents a timestamp split into milliseconds and
nanoseconds for JSON serialization
type PreciseTimestamp struct {
	Milliseconds int64 `json:"milliseconds"`
	Nanoseconds  int32 `json:"nanoseconds"`
}
`
I had to thread this functionality through on the front-end in trace
summaries, spans, and events as well, and get rid of the code that
parsed string-based timestamps, so that accounts for good chunk of the
changes.

Finally, I added a route for logs behind a feature flag to print out log
json. It's a bit of a hack with our current front end, but it's just to
confirm that it works:

<img width="589" alt="isBigIsHeavyIsWood"
src="https://github.com/user-attachments/assets/de56cc30-486d-43ad-91f3-76c246759f99"
/>


![Gemini_Generated_Image_arcip5arcip5arci](https://github.com/user-attachments/assets/135db0e0-5e0f-4bab-a53a-f02f62a4e13d)